[avani]Refactor.  from float datatype to double.

+++ src/main/java/com/thoughtworks/bootcamp/shoppingCart/ShoppingCart.java
@@ -1,21 +1,19 @@
 package com.thoughtworks.bootcamp.shoppingCart;
 
-import java.util.ArrayList;
-
 public class ShoppingCart {
     private String apple;
-    private float price;
+    private double price;
     private int quantity;
 
-    public ShoppingCart(String apple, float price, int quantity) {
+    public ShoppingCart(String apple, double price, int quantity) {
         this.apple = apple;
         this.price = price;
         this.quantity = quantity;
     }
 
-    public float totalPrice(float price) {
-        float totalPrice = price * getQuantity();
-        return totalPrice;
+    public double totalPrice(double price) {
+        double totalPrice = price * getQuantity();
+        return  totalPrice;
     }
 
     public int getQuantity() {
