+++ src/test/java/com/thoughtworks/bootcamp/shoppingCart/ShoppingCartTest.java
@@ -3,27 +3,27 @@ package com.thoughtworks.bootcamp.shoppingCart;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
 
 public class ShoppingCartTest {
 
     @Test
     void givenEmptyShoppingCart_whenAddOneApple_thenShouldReturnOneApplePrice() {
-        ShoppingCart cart = new ShoppingCart("apple", (float) 0.99, 1);
-        float oneApplePrice = (float) 0.99;
+        ShoppingCart cart = new ShoppingCart("apple", 0.99, 1);
+        double oneApplePrice = 0.99;
 
-        Assertions.assertEquals((float) 0.99, cart.totalPrice(oneApplePrice));
+        Assertions.assertEquals(0.99, cart.totalPrice(oneApplePrice));
 
     }
 
     @Test
     void givenShoppingCart_whenAppleQuantityIsTwo_thenShouldReturnPriceOnePointNineEight() {
 
-        ShoppingCart shoppingCart = new ShoppingCart("apple", (float) 0.99, 2);
+        ShoppingCart shoppingCart = new ShoppingCart("apple", 0.99, 2);
 
-        float applePrice = (float) 0.99;
-        Assertions.assertEquals((float)1.98, shoppingCart.totalPrice(applePrice));
+        double applePrice = 0.99;
+        Assertions.assertEquals(1.98, shoppingCart.totalPrice(applePrice));
 
     }
+
 
 }
